refactor: centralize gml identifier metadata access
