+++ src/semantic/src/resources/bundled-resources.js
@@ -1,19 +1,8 @@
-import { createRequire } from "node:module";
-import { fileURLToPath } from "node:url";
-
-const require = createRequire(import.meta.url);
-
-export const GML_IDENTIFIER_METADATA_URL = new URL(
-    "../../../../resources/gml-identifiers.json",
-    import.meta.url
-);
-export const GML_IDENTIFIER_METADATA_PATH = fileURLToPath(
-    GML_IDENTIFIER_METADATA_URL
-);
-
-export function loadBundledIdentifierMetadata() {
-    return require(GML_IDENTIFIER_METADATA_PATH);
-}
+export {
+    GML_IDENTIFIER_METADATA_PATH,
+    GML_IDENTIFIER_METADATA_URL,
+    loadBundledIdentifierMetadata
+} from "@prettier-plugin-gml/shared/resources/gml-identifiers.js";
 export {
     FEATHER_METADATA_PATH,
     loadBundledFeatherMetadata
