+++ src/shared/src/resources/index.js
@@ -6,3 +6,10 @@ export {
     getFeatherMetadata,
     loadBundledFeatherMetadata
 } from "./feather-metadata.js";
+export {
+    clearIdentifierMetadataCache,
+    getIdentifierMetadata,
+    GML_IDENTIFIER_METADATA_PATH,
+    GML_IDENTIFIER_METADATA_URL,
+    loadBundledIdentifierMetadata
+} from "./gml-identifiers.js";
